fix: close model fallback review findings
