+++ scripts/agent-evidence
@@ -101,8 +101,12 @@ function parseSelection(argv) {
     }
   }
 
+  const requestsModelConsultation = requestedOnly.includes("model-consultation");
   for (let index = 0; index < requestedOnly.length; index += 1) {
     requestedOnly[index] = laneAliases.get(requestedOnly[index]) || requestedOnly[index];
+  }
+  if (requestsModelConsultation && !requestedOnly.includes("association-audit")) {
+    requestedOnly.push("association-audit");
   }
 
   const known = new Set(allLanes.map((lane) => lane.id));
