+++ src/struct/index.ts
@@ -1,5 +1,9 @@
 export * from './types'
 export * from './ids'
+export {
+  validateModelConsultationReceipt,
+  type ModelFallbackReceipt,
+} from './model-consultation-receipt'
 export * from './reading-order'
 export * from './recovery'
 export * from './from-reconstruction'
